Migrate regressions from 3000-3501
