+++ spacy/tests/doc/test_span.py
@@ -69,6 +69,18 @@ def test_issue1612(en_tokenizer):
     assert span.orth_ == span.text
 
 
+@pytest.mark.issue(3199)
+def test_issue3199():
+    """Test that Span.noun_chunks works correctly if no noun chunks iterator
+    is available. To make this test future-proof, we're constructing a Doc
+    with a new Vocab here and a parse tree to make sure the noun chunks run.
+    """
+    words = ["This", "is", "a", "sentence"]
+    doc = Doc(Vocab(), words=words, heads=[0] * len(words), deps=["dep"] * len(words))
+    with pytest.raises(NotImplementedError):
+        list(doc[0:3].noun_chunks)
+
+
 @pytest.mark.parametrize(
     "i_sent,i,j,text",
     [
