+++ spacy/tests/test_displacy.py
@@ -1,4 +1,6 @@
+import numpy
 import pytest
+
 from spacy import displacy
 from spacy.displacy.render import DependencyRenderer, EntityRenderer
 from spacy.tokens import Span, Doc
@@ -26,6 +28,18 @@ def test_issue2728(en_vocab):
     doc.ents = [Span(doc, 1, 2, label="TEST")]
     html = displacy.render(doc, style="ent")
     assert "&lt;RELEASE&gt;" in html
+
+
+@pytest.mark.issue(3288)
+def test_issue3288(en_vocab):
+    """Test that retokenization works correctly via displaCy when punctuation
+    is merged onto the preceeding token and tensor is resized."""
+    words = ["Hello", "World", "!", "When", "is", "this", "breaking", "?"]
+    heads = [1, 1, 1, 4, 4, 6, 4, 4]
+    deps = ["intj", "ROOT", "punct", "advmod", "ROOT", "det", "nsubj", "punct"]
+    doc = Doc(en_vocab, words=words, heads=heads, deps=deps)
+    doc.tensor = numpy.zeros((len(words), 96), dtype="float32")
+    displacy.render(doc)
 
 
 def test_displacy_parse_ents(en_vocab):
